C2_Server/requirements.txt
flask==3.0.2
flask-login==0.6.3
werkzeug==3.0.1
cryptography==42.0.5
requests==2.31.0
dnspython==2.5.0
pycryptodomex==3.20.0
pywin32==306; sys_platform == 'win32'
psutil==5.9.8
python-dotenv==1.0.1
pefile==2023.2.7
capstone==5.0.1
keystone-engine==0.9.2
flask-wtf==1.2.1
flask-sqlalchemy==3.1.1
bcrypt==4.1.2 